14/05/2021

+++ src/main/resources/initBD.sql
@@ -1,13 +1,13 @@
 CREATE TABLE product
 (
     id          INTEGER,
-    category    VARCHAR(255)     NOT NULL,
-    name        VARCHAR(255)     NOT NULL,
-    price       DOUBLE DEFAULT 0 NOT NULL,
-    discount    DOUBLE DEFAULT 0 NOT NULL,
-    totalVolume DOUBLE DEFAULT 0 NOT NULL,
-    actualPrice DOUBLE DEFAULT 0 NOT NULL,
-    data        DATETIME         NOT NULL
+    category    VARCHAR(255)            NOT NULL,
+    name        VARCHAR(255)            NOT NULL,
+    price       DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    discount    DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    totalVolume DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    actualPrice DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    data        DATETIME                NOT NULL
 );
 
 CREATE TABLE person
@@ -20,20 +20,20 @@ CREATE TABLE person
 CREATE TABLE productOfDelete
 (
     id          INTEGER,
-    category    VARCHAR(255)     NOT NULL,
-    name        VARCHAR(255)     NOT NULL,
-    price       DOUBLE DEFAULT 0 NOT NULL,
-    discount    DOUBLE DEFAULT 0 NOT NULL,
-    actualPrice DOUBLE DEFAULT 0 NOT NULL,
-    totalVolume DOUBLE DEFAULT 0 NOT NULL,
-    data        DATETIME         NOT NULL
+    category    VARCHAR(255)            NOT NULL,
+    name        VARCHAR(255)            NOT NULL,
+    price       DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    discount    DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    actualPrice DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    totalVolume DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    data        DATETIME                NOT NULL
 );
 
 CREATE TABLE buyproduct
 (
     id          INTEGER,
-    name        VARCHAR(255)     NOT NULL,
-    actualPrice DOUBLE DEFAULT 0 NOT NULL,
-    quantity    DOUBLE DEFAULT 0 NOT NULL,
-    totalPrice  DOUBLE DEFAULT 0 NOT NULL
+    name        VARCHAR(255)            NOT NULL,
+    actualPrice DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    quantity    DOUBLE(10, 3) DEFAULT 0 NOT NULL,
+    totalPrice  DOUBLE(10, 3) DEFAULT 0 NOT NULL
 )
